Rename weapon_mom_smg -> weapon_mom_machinegun
Updated recoil to be less intense while moving
