+++ mp/src/game/client/client_momentum.vpc
@@ -300,8 +300,8 @@ $Project "Client (Momentum)"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_stickybomblauncher.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_shotgun.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_shotgun.cpp"
-                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_smg.h"
-                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_smg.cpp"
+                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_machinegun.h"
+                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_machinegun.cpp"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_sniper.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_sniper.cpp"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_grenade.h"
