+++ mp/src/game/server/server_momentum.vpc
@@ -111,8 +111,8 @@ $Project "Server (Momentum)"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_stickybomblauncher.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_shotgun.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_shotgun.cpp"
-                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_smg.h"
-                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_smg.cpp"
+                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_machinegun.h"
+                $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_machinegun.cpp"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_sniper.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_sniper.cpp"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_mom_grenade.h"
