+++ mp/src/game/shared/momentum/weapon/weapon_mom_machinegun.cpp
@@ -1,6 +1,6 @@
 #include "cbase.h"
 
-#include "weapon_mom_smg.h"
+#include "weapon_mom_machinegun.h"
 #include "mom_player_shared.h"
 
 #include "tier0/memdbgon.h"
@@ -38,9 +38,9 @@ void CMomentumMachinegun::PrimaryAttack()
 
     // Kick the gun based on the state of the player.
     if (!FBitSet(pPlayer->GetFlags(), FL_ONGROUND))
-        pPlayer->KickBack(0.9, 0.45, 0.35, 0.04, 5.25, 3.5, 4);
+        pPlayer->KickBack(.33, 0.23, 0.18, 0.020, 3.7, 2.1, 4.5);
     else if (pPlayer->GetAbsVelocity().Length2D() > 5)
-        pPlayer->KickBack(0.45, 0.3, 0.2, 0.0275, 4, 2.25, 7);
+        pPlayer->KickBack(0.40, 0.25, 0.2, 0.0260, 4, 2.17, 4.9);
     else if (FBitSet(pPlayer->GetFlags(), FL_DUCKING))
         pPlayer->KickBack(0.275, 0.2, 0.125, 0.02, 3, 1, 9);
     else
