+++ mp/src/game/shared/momentum/weapon/weapon_mom_machinegun.h
